Feat: 칼럼 리스트 컴포넌트 분리
스크롤 관련해서 반응형에서 정상적으로 동작하도록 수정하고, 컴포넌트를 분리했습니다.
