+++ src/pages/dashboard.{dashboardid}/components/ColumnList/ColumnList.module.scss
@@ -1,0 +1,10 @@
+@import '../../../../styles/utils.scss';
+
+.addColumnButtonContainer {
+  width: auto;
+  @include desktop {
+    padding-top: 4.25rem;
+    padding-left: 1.25rem;
+    padding-right: 3rem;
+  }
+}
